requirements.txt
alembic==1.8.1
click==8.1.3
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0
importlib-metadata==5.0.0
importlib-resources==5.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
SQLAlchemy==1.4.42
Werkzeug==2.2.2
zipp==3.9.0
python-dotenv
flask-smorest
flask-jwt-extended
passlib
pymysql==1.0.2
cryptography
flask-migrate
requests
rq
faker
injector
flask_restful
flask-uploads
